requirements.txt
PyQt5>=5.15
pytesseract
pillow
